feat: Exhibition 엔티티 필드 추가 및 폴더 구조 리팩토링
