+++ src/main/java/com/promesa/promesa/domain/exhibition/dao/ExhibitionRepository.java
@@ -1,6 +1,6 @@
-package com.promesa.promesa.domain.home.dao;
+package com.promesa.promesa.domain.exhibition.dao;
 
-import com.promesa.promesa.domain.item.domain.Exhibition;
+import com.promesa.promesa.domain.exhibition.domain.Exhibition;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ExhibitionRepository extends JpaRepository<Exhibition, Long> {
