+++ src/main/java/com/promesa/promesa/domain/exhibition/domain/ExhibitionStatus.java
@@ -1,0 +1,5 @@
+package com.promesa.promesa.domain.exhibition.domain;
+
+public enum ExhibitionStatus {
+    ONGOING, ENDED
+}
