+++ src/main/java/com/promesa/promesa/domain/exhibition/exception/ExhibitionErrorCode.java
@@ -1,4 +1,4 @@
-package com.promesa.promesa.domain.home.exception;
+package com.promesa.promesa.domain.exhibition.exception;
 
 import com.promesa.promesa.common.dto.ErrorReason;
 import com.promesa.promesa.common.exception.BaseErrorCode;
@@ -9,7 +9,7 @@ import static com.promesa.promesa.common.consts.PromesaStatic.NOT_FOUND;
 
 @Getter
 @AllArgsConstructor
-public enum HomeErrorCode implements BaseErrorCode {
+public enum ExhibitionErrorCode implements BaseErrorCode {
     EXHIBITION_NOT_FOUND(NOT_FOUND, "Exhibition_404", "해당 기획전을 찾을 수 없습니다.");
 
     private Integer status;
