+++ src/main/java/com/promesa/promesa/domain/exhibition/exception/ExhibitionNotFoundException.java
@@ -1,10 +1,10 @@
-package com.promesa.promesa.domain.home.exception;
+package com.promesa.promesa.domain.exhibition.exception;
 
 import com.promesa.promesa.common.exception.PromesaCodeException;
 
 public class ExhibitionNotFoundException extends PromesaCodeException {
     public static final PromesaCodeException EXCEPTION = new ExhibitionNotFoundException();
     private ExhibitionNotFoundException() {
-        super(HomeErrorCode.EXHIBITION_NOT_FOUND);
+        super(ExhibitionErrorCode.EXHIBITION_NOT_FOUND);
     }
 }
